% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.stats-package.R
\docType{package}
\name{learning.stats-package}
\alias{learning.stats-package}
\alias{learn.stats}
\title{Learning Statistics}
\description{
\code{learn.stats} provides a range of tools for learning R.
}
